Short if route params cahnge
Hey, could you please check this? I've added a check that the go function is executed too, if the params change. Current bug is, that if the starting route has a param, that the route is not recognised as being active and executed again, if the user clicks on the same s-ref.